// src/department/dto/response/response-department.dto.ts
export class ResponseDepartmentDto {
    departmentid: number;
    departmentname: string;
    customerid: number;
    facilityid: number;
    location: string;
    is_active: boolean;
    date_created: Date;
    date_updated: Date;
    is_deleted: boolean;
    created_by: number;
    updated_by: number;
}